nits from PR review
