+++ app/src/main/java/com/bottlerocketstudios/brarchitecture/ui/pullrequests/PullRequestViewModel.kt
@@ -1,14 +1,12 @@
 package com.bottlerocketstudios.brarchitecture.ui.pullrequests
 
-import androidx.lifecycle.viewModelScope
+import com.bottlerocketstudios.brarchitecture.R
 import com.bottlerocketstudios.brarchitecture.data.repository.BitbucketRepository
-import com.bottlerocketstudios.brarchitecture.domain.models.Status
 import com.bottlerocketstudios.brarchitecture.ui.BaseViewModel
 import com.bottlerocketstudios.compose.pullrequest.PullRequestItemState
 import com.bottlerocketstudios.compose.util.asMutableState
 import com.bottlerocketstudios.compose.util.formattedUpdateTime
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
 import org.koin.core.component.inject
 import java.time.Clock
 
@@ -18,7 +16,7 @@ class PullRequestViewModel : BaseViewModel() {
     private val repo: BitbucketRepository by inject()
     private val clock by inject<Clock>()
 
-    val pullRequestRequestList = repo.pullRequests.map {
+    val pullRequestList = repo.pullRequests.map {
         it.map { dto ->
             PullRequestItemState(
                 prName = dto.title.orEmpty().asMutableState(),
@@ -32,16 +30,8 @@ class PullRequestViewModel : BaseViewModel() {
 
     // Init logic
     init {
-        viewModelScope.launch(dispatcherProvider.IO) {
-            val results = repo.getPullRequests()
-            when (results) {
-                is Status.Success -> {}
-                is Status.Failure -> {
-                    if (results is Status.Failure.GeneralFailure) {
-                        toaster.toast(results.message)
-                    }
-                }
-            }
+        launchIO {
+            repo.getPullRequests().handlingErrors(R.string.pull_request_error) {}
         }
     }
 }
