+++ compose/src/main/java/com/bottlerocketstudios/compose/pullrequest/PullRequestItemCard.kt
@@ -9,32 +9,22 @@ import androidx.compose.material.Card
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.bottlerocketstudios.compose.R
 import com.bottlerocketstudios.compose.resources.Colors
 import com.bottlerocketstudios.compose.resources.Dimens
 import com.bottlerocketstudios.compose.resources.greyish_brown
-import com.bottlerocketstudios.compose.resources.sea_foam
 import com.bottlerocketstudios.compose.util.Preview
 import com.bottlerocketstudios.compose.util.PreviewComposable
+import com.bottlerocketstudios.compose.util.ResponsiveText
 import com.bottlerocketstudios.compose.util.asMutableState
-
-private const val TEXT_SCALE_REDUCTION_INTERVAL = 0.9f
 
 @Composable
 fun PullRequestItemCard(state: PullRequestItemState) {
@@ -104,38 +94,4 @@ fun PullRequestsItemPreview() {
             )
         )
     }
-}
-
-@Composable
-fun ResponsiveText(
-    modifier: Modifier = Modifier,
-    text: String,
-    color: Color = sea_foam,
-    textAlign: TextAlign = TextAlign.Start,
-    textStyle: TextStyle,
-    targetTextSizeHeight: TextUnit = textStyle.fontSize,
-    maxLines: Int = 2,
-) {
-    var textSize by remember { mutableStateOf(targetTextSizeHeight) }
-
-    Text(
-        modifier = modifier,
-        text = text,
-        color = color,
-        textAlign = textAlign,
-        fontSize = textSize,
-        fontFamily = textStyle.fontFamily,
-        fontStyle = textStyle.fontStyle,
-        fontWeight = textStyle.fontWeight,
-        lineHeight = textStyle.lineHeight,
-        maxLines = maxLines,
-        overflow = TextOverflow.Ellipsis,
-        onTextLayout = { textLayoutResult ->
-            val maxCurrentLineIndex: Int = textLayoutResult.lineCount - 1
-
-            if (textLayoutResult.isLineEllipsized(maxCurrentLineIndex)) {
-                textSize = textSize.times(TEXT_SCALE_REDUCTION_INTERVAL)
-            }
-        },
-    )
 }
